requirements.txt
fastapi
uvicorn
sentence-transformers
transformers
torch
torchvision
torchaudio
pillow
pymupdf
openai-whisper
python-multipart
pydantic
requests
python-dotenv
cloudinary
moviepy
yt-dlp
beautifulsoup4
playwright
html2text
fpdf2
python-docx
google-api-python-client
